Add files via upload
I have added 4 seperate boards, each with their own 2d array. This way when player1 is choosing a place to see if there is a battleship, we can use that same location to check for player2. The click event tells us where we are in the relation to the window."